refactor: continue daemon and telemetry decomposition
