+++ internal/telemetry/usage_view.go
@@ -7,7 +7,6 @@ import (
 	"sort"
 	"strings"
 	"time"
-	"unicode"
 
 	"github.com/janekbaraniewski/openusage/internal/core"
 	"github.com/samber/lo"
@@ -1637,220 +1636,3 @@ func sortedSeriesFromByDay(byDay map[string]float64) []core.TimePoint {
 }
 
 // parseMCPToolName extracts server and function from an MCP tool name.
-// Raw tool names use double underscores: mcp__server__function.
-// Returns ("", "", false) for non-MCP tools.
-// parseMCPToolName extracts server and function from an MCP tool name.
-// Supports two formats:
-//   - Canonical: "mcp__server__function" (double underscores, from Claude Code and normalized Cursor)
-//   - Legacy:    "server-function (mcp)" or "user-server-function (mcp)" (old Cursor data)
-//
-// Returns ("", "", false) for non-MCP tools.
-func parseMCPToolName(raw string) (server, function string, ok bool) {
-	raw = strings.ToLower(strings.TrimSpace(raw))
-
-	// Canonical format: mcp__server__function
-	if strings.HasPrefix(raw, "mcp__") {
-		rest := raw[5:]
-		idx := strings.Index(rest, "__")
-		if idx < 0 {
-			return rest, "", true
-		}
-		return rest[:idx], rest[idx+2:], true
-	}
-
-	// Copilot legacy wrapper format: "<server>_mcp_server_<function>".
-	if strings.Contains(raw, "_mcp_server_") {
-		parts := strings.SplitN(raw, "_mcp_server_", 2)
-		server = sanitizeMCPToolSegment(parts[0])
-		function = sanitizeMCPToolSegment(parts[1])
-		if server != "" && function != "" {
-			return server, function, true
-		}
-	}
-
-	// Copilot legacy wrapper format variant: "<server>-mcp-server-<function>".
-	if strings.Contains(raw, "-mcp-server-") {
-		parts := strings.SplitN(raw, "-mcp-server-", 2)
-		server = sanitizeMCPToolSegment(parts[0])
-		function = sanitizeMCPToolSegment(parts[1])
-		if server != "" && function != "" {
-			return server, function, true
-		}
-	}
-
-	// Legacy format: "something (mcp)" from old Cursor data.
-	if strings.HasSuffix(raw, " (mcp)") {
-		body := strings.TrimSuffix(raw, " (mcp)")
-		body = strings.TrimSpace(body)
-		if body == "" {
-			return "", "", false
-		}
-
-		// Strip "user-" prefix if present.
-		body = strings.TrimPrefix(body, "user-")
-
-		// Try to extract server from "server-function" format.
-		// e.g., "kubernetes-pods_log" → server=kubernetes, function=pods_log
-		// e.g., "gcp-gcloud-run_gcloud_command" → server=gcp-gcloud, function=run_gcloud_command
-		// Heuristic: the function part typically contains underscores, so split on the
-		// last hyphen that precedes an underscore-containing segment.
-		if idx := findServerFunctionSplit(body); idx > 0 {
-			return body[:idx], body[idx+1:], true
-		}
-
-		// No clear server-function split — treat whole body as function with unknown server.
-		return "other", body, true
-	}
-
-	return "", "", false
-}
-
-func sanitizeMCPToolSegment(raw string) string {
-	raw = strings.ToLower(strings.TrimSpace(raw))
-	if raw == "" {
-		return ""
-	}
-	var b strings.Builder
-	b.Grow(len(raw))
-	lastUnderscore := false
-	for _, r := range raw {
-		if unicode.IsLetter(r) || unicode.IsDigit(r) {
-			b.WriteRune(r)
-			lastUnderscore = false
-			continue
-		}
-		if !lastUnderscore {
-			b.WriteRune('_')
-			lastUnderscore = true
-		}
-	}
-	return strings.Trim(b.String(), "_")
-}
-
-// findServerFunctionSplit finds the best hyphen position to split "server-function"
-// in a Cursor MCP tool name. The function name typically contains underscores
-// (e.g., "pods_list", "search_docs") while server names use hyphens.
-// Strategy: find the last hyphen where the part AFTER it contains an underscore.
-// This handles multi-segment server names like "gcp-gcloud" or "runai-docs".
-func findServerFunctionSplit(s string) int {
-	bestIdx := -1
-	for i := 0; i < len(s); i++ {
-		if s[i] == '-' {
-			rest := s[i+1:]
-			if strings.Contains(rest, "_") {
-				bestIdx = i
-			}
-		}
-	}
-	if bestIdx > 0 {
-		return bestIdx
-	}
-
-	// No underscore-based split found. Fall back to first hyphen if no more hyphens after.
-	// e.g., "kubernetes-pods_log" or "smart-query"
-	if idx := strings.Index(s, "-"); idx > 0 {
-		return idx
-	}
-	return -1
-}
-
-func buildMCPAgg(tools []telemetryToolAgg) []telemetryMCPServerAgg {
-	type serverData struct {
-		calls   float64
-		calls1d float64
-		funcs   map[string]*telemetryMCPFunctionAgg
-	}
-	servers := make(map[string]*serverData)
-
-	for _, tool := range tools {
-		server, function, ok := parseMCPToolName(tool.Tool)
-		if !ok || server == "" {
-			continue
-		}
-		sd, exists := servers[server]
-		if !exists {
-			sd = &serverData{funcs: make(map[string]*telemetryMCPFunctionAgg)}
-			servers[server] = sd
-		}
-		sd.calls += tool.Calls
-		sd.calls1d += tool.Calls1d
-		if function != "" {
-			if f, ok := sd.funcs[function]; ok {
-				f.Calls += tool.Calls
-				f.Calls1d += tool.Calls1d
-			} else {
-				sd.funcs[function] = &telemetryMCPFunctionAgg{
-					Function: function,
-					Calls:    tool.Calls,
-					Calls1d:  tool.Calls1d,
-				}
-			}
-		}
-	}
-
-	result := make([]telemetryMCPServerAgg, 0, len(servers))
-	for name, sd := range servers {
-		var funcs []telemetryMCPFunctionAgg
-		for _, f := range sd.funcs {
-			funcs = append(funcs, *f)
-		}
-		sort.Slice(funcs, func(i, j int) bool {
-			if funcs[i].Calls != funcs[j].Calls {
-				return funcs[i].Calls > funcs[j].Calls
-			}
-			return funcs[i].Function < funcs[j].Function
-		})
-		result = append(result, telemetryMCPServerAgg{
-			Server:    name,
-			Calls:     sd.calls,
-			Calls1d:   sd.calls1d,
-			Functions: funcs,
-		})
-	}
-	sort.Slice(result, func(i, j int) bool {
-		if result[i].Calls != result[j].Calls {
-			return result[i].Calls > result[j].Calls
-		}
-		return result[i].Server < result[j].Server
-	})
-	return result
-}
-
-// deleteByPrefixes removes all entries from a string-keyed map whose key
-// matches any of the given prefixes. Works with any map[string]V.
-func deleteByPrefixes[V any](m map[string]V, prefixes []string) {
-	for key := range m {
-		for _, p := range prefixes {
-			if strings.HasPrefix(key, p) {
-				delete(m, key)
-				break
-			}
-		}
-	}
-}
-
-func sanitizeMetricID(raw string) string {
-	raw = strings.TrimSpace(strings.ToLower(raw))
-	if raw == "" {
-		return "unknown"
-	}
-	var b strings.Builder
-	b.Grow(len(raw))
-	lastUnderscore := false
-	for _, r := range raw {
-		if unicode.IsLetter(r) || unicode.IsDigit(r) {
-			b.WriteRune(r)
-			lastUnderscore = false
-			continue
-		}
-		if !lastUnderscore {
-			b.WriteRune('_')
-			lastUnderscore = true
-		}
-	}
-	out := strings.Trim(b.String(), "_")
-	if out == "" {
-		return "unknown"
-	}
-	return out
-}
